requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
numpy
scikit-learn
openai
python-dotenv
aiofiles
pydantic
pypdf
python-docx
